requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.1
pandas>=2.0.3
openpyxl>=3.1.2
streamlit>=1.24.0
chromadb>=0.4.13
matplotlib>=3.7.2
python-dotenv>=1.0.0
chromadb>=0.4.13